Refactor entitlements files for macOS Runner
