+++ apps/dashboard/lib/src/features/auth_gate/presentation/auth_gate.dart
@@ -1,0 +1,21 @@
+import 'package:dashboard/src/features/sign_up/presentation/sign_up_page.dart';
+import 'package:dashboard/src/features/workflow/presentation/workflow_page.dart';
+import 'package:firebase_auth/firebase_auth.dart';
+import 'package:flutter/material.dart';
+
+class AuthGate extends StatelessWidget {
+  const AuthGate({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return StreamBuilder(
+      stream: FirebaseAuth.instance.authStateChanges(),
+      builder: (context, snapshot) {
+        if (snapshot.hasData) {
+          return const WorkflowPage();
+        }
+        return const SignUpPage();
+      },
+    );
+  }
+}
